replace all courses int to long type and remove CourseId in CourseReq

+++ src/StudyThink.Service/DTOs/Courses/Course/CourseUpdateDto.cs
@@ -4,10 +4,10 @@ namespace StudyThink.Service.DTOs.Courses.Course;
 
 public class CourseUpdateDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
     public float Price { get; set; }
     public IFormFile? ImagePath { get; set; }
     public float TotalPrice { get; set; }
@@ -15,6 +15,6 @@ public class CourseUpdateDto
     public float Duration { get; set; }
     public string Language { get; set; }
     public float DiscountPrice { get; set; }
-    public int CourseReqId { get; set; }
+    public long CourseReqId { get; set; }
 
 }
